requirements.txt
pytest~=8.3.4
selenium~=4.27.1
allure-python-commons~=2.13.5
Faker~=34.0.2